+++ gradle.properties
@@ -13,5 +13,5 @@ fabric_version=0.104.2+1.21.2
 
 # Mod Properties
 mod_version = 3.0.0
-maven_group = tk.estecka.invarpaint
+maven_group = fr.estecka.invarpaint
 archives_base_name = invariable-paintings
