+++ src/client/resources/invarpaint-pick.mixins.json
@@ -1,6 +1,6 @@
 {
 	"required": true,
-	"package": "tk.estecka.invarpaint.pick.mixin",
+	"package": "fr.estecka.invarpaint.pick.mixin",
 	"compatibilityLevel": "JAVA_21",
 	"mixins": [
 		"ItemGroupsMixin",
